Abstract Interpretation

PDDL.jl supports abstract interpretation of PDDL domains, the semantics of which can also be compiled using the PDDL.jl compiler. This functionality is exposed by the abstracted and abstractstate functions:

PDDL.abstracted — Function

abstracted(domain; options...)

Construct an abstract domain from a concrete domain. See PDDL.AbstractInterpreter for the list of options.

abstracted(domain, state; options...)
abstracted(domain, problem; options...)

Construct an abstract domain and state from a concrete domain and state. A problem can be provided instead of a state. See PDDL.AbstractInterpreter for the list of options.

PDDL.abstractstate — Function

abstractstate(domain, state)

Construct a state in an abstract domain from a concrete `state.

The behavior of the abstract interpreter can be customized by specifying the Julia type used to represent abstract values for a particular fluent or PDDL type:

Missing docstring.

Missing docstring for PDDL.AbstractInterpeter. Check Documenter's build log for details.

Abstract Values and Types

Abstract interpretation requires each concrete value to be mapped to an abstract value of a particular type, which represents an (over-approximation) of the set of the possible values reachable after a series of actions have been executed. By default, Boolean values (i.e. predicates) are mapped to the BooleanAbs abstraction, while scalar numbers (corresponding to PDDL types like integer, number and numeric) are mapped to the IntervalAbs abstraction. Other types of values may use the SetAbs abstraction.

PDDL.BooleanAbs — Type

BooleanAbs

Belnap logic abstraction for Boolean values. In addition to true, false and missing, the abstract value both can be used to represent when a predicate is both false and true.

PDDL.IntervalAbs — Type

IntervalAbs(lo::Real, hi::Real)

Interval abstraction for real-valued numbers.

PDDL.SetAbs — Type

SetAbs(xs...)

Set abstraction for arbitrary objects.