Parser and Writer

PDDL.jl supports both parsing and writing of PDDL files and strings. In addition, the parser is designed to be extensible, allowing variants or extensions of PDDL to be easily supported.

General Parsing

The PDDL.Parser submodule contains all functionality related to parsing PDDL strings and loading of PDDL files. To parse a string in PDDL, use the macro @pddl or the function parse_pddl. Both of these return a list of parsed results if multiple strings are provided.

PDDL.Parser.@pddl — Macro

Parse string(s) to PDDL construct.

PDDL.Parser.parse_pddl — Function

Parse to PDDL structure based on initial keyword.

Below we use @pddl to parse a sequence of predicates, and use parse_pddl to parse a PDDL axiom (a.k.a. derived predicate):

julia> @pddl("(on a b)", "(on b c)")
2-element Vector{Compound}:
 on(a, b)
 on(b, c)
julia> parse_pddl("(:derived (handempty) (forall (?x) (not (holding ?x))))")
handempty <<= forall(object(X), not(holding(X)))

In addition, there exists a string macro pddl"...", which is ideal for parsing single string literals:

julia> pddl"(on a b)"
on(a, b)

Parsing Domains and Problems

To parse domains and problems specified as PDDL strings, use parse_domain and parse_problem.

PDDL.Parser.parse_domain — Function

Parse PDDL domain description.

PDDL.Parser.parse_problem — Function

Parse PDDL problem description.

To load domains or problems from a file, use load_domain and load_problem.

PDDL.Parser.load_domain — Function

load_domain(path::AbstractString)
load_domain(io::IO)

Load PDDL domain from specified path or IO object.

PDDL.Parser.load_problem — Function

load_problem(path::AbstractString)
load_problem(io::IO)

Load PDDL problem from specified path or IO object.

Extending the Parser

The parser can be extended to handle new PDDL constructs using the following macros:

PDDL.Parser.@add_top_level — Macro

@add_top_level(name, f)

Register f as a top-level parser for PDDL descriptions (e.g domains, problems).

PDDL.Parser.@add_header_field — Macro

@add_header_field(desc, fieldname, f)

Register f as a parser for a header field in a PDDL description.

PDDL.Parser.@add_body_field — Macro

@add_body_field(desc, fieldname, f)

Register f as a parser for a body field in a PDDL description.

General Writing

The PDDL.Writer submodule contains all functionality related to writing PDDL strings and saving of PDDL files. To write a string in PDDL syntax, use the function write_pddl.

PDDL.Writer.write_pddl — Function

Write to string in PDDL syntax.

Below we use write_pddl to write out an Action from the Blocksworld domain.

julia> write_pddl(PDDL.get_action(domain, :stack)) |> print
(:action stack
 :parameters (?x ?y - block)
 :precondition (and (holding ?x) (clear ?y) (not (= ?x ?y)))
 :effect (and (not (holding ?x)) (not (clear ?y)) (clear ?x) (handempty) (on ?x ?y)))

Writing Domains and Problems

To write domains and problem as PDDL strings, use write_domain and write_problem.

PDDL.Writer.write_domain — Function

Write domain in PDDL syntax.

PDDL.Writer.write_problem — Function

Write problem in PDDL syntax.

To save domains or problems as text files to a path, use save_domain and save_problem.

PDDL.Writer.save_domain — Function

save_domain(path::String, domain::Domain)

Save PDDL domain to specified path.

PDDL.Writer.save_problem — Function

save_problem(path::String, problem::Problem)

Save PDDL problem to specified path.